Define script that can detect when the player is grounded
The TouchingDirections script hanled detecting if the player is grounded or not.

This commit also fixes a bug where the Player instance in GameplayScene actually had 2 capsule colliders instead of one. This has now been fixed. And with that fix the prefab Player has the capsule collider properly sized now.