Knight plays attack animation when it detects Player
Adds a 'scan box' to the knight. If the Knight detects the player, the attack is triggered.
While attacking, the knight doesn't move (but keeps momentum slightly)

Also extended the disabling of movement of the Knight to Death and Hit animations.

Death and Hit animations added to the Knight enemy currently not triggered by anything yet.