Implement hit-stun and knockback for the Player character
When the Knight Enemy attacks the Player, the Player briefly cannot move and suffers knockback.

The hit registration is imperfect though. A single swing from the Knight Enemy can register multiple hits on the Player, which is unwanted.
Even after increasing the invicibility timer for the Player on hit to longer than the Knight Enemy Attack hitbox duration, multi-hitting still happens.